requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beartype==0.22.9
beautifulsoup4==4.14.3
boolean.py==5.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
distro==1.9.0
fastapi==0.135.1
feedparser==6.0.12
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.18
langchain-groq==1.1.2
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.10
langsmith==0.7.16
license-expression==30.4.4
numpy==2.4.3
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.1
ply==3.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
rdflib==7.6.0
requests==2.32.5
requests-toolbelt==1.0.0
semantic-version==2.10.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
spdx-tools==0.8.3
sqlparse==0.5.5
starlette==0.52.1
structlog==25.5.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
uritools==6.0.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
xmltodict==1.0.4
xxhash==3.6.0
zstandard==0.25.0
